Added skelton csv view to Team Schedule action. (#14)
