+++ src/AppBundle/Resources/config/services/app.yml
@@ -20,6 +20,14 @@ services:
            - ['setBaseTemplate', ['@app_base_template']]
            - ['setProject', ['%app_project%']]
 
+    abstract_view:
+        abstract: true
+        calls:
+           - ['setRouter',  ['@router']]
+           - ['setSecurity',['@security.token_storage','@security.authorization_checker']]
+           - ['setBaseTemplate', ['@app_base_template']]
+           - ['setProject', ['%app_project%']]
+
     app_action_abstract_controller:
         abstract: true
         calls: [[setContainer, ['@service_container']]]
